chore: add shell.nix for nix development environment (#2)
Do note that the linked issue is only partially incomplete. The reason
is that mask-hx is primarily imperative, unlike NixOS itself, which
provides the Nix package manager out of the box that allows setting up
development environments the same way as mask-hx, but declaratively.

As such, NixOS support won't be available, even though I'm using NixOS
now. Either use a Nix development environment, a Flake, or run mask-hx
in a container.