feat: reimplement check functionality in the binary
Some of mask-hx is available for testing, but most subcommands aren't implemented.

This also changes some parts of the general API to work fairly better, and removes
certain unnecessary pieces of data, especially CommandResult, which increases the total
lines of code. It also fixes how getting a path using HaxeVersion would return an
entirely new path instead of the buffer.