refactor: reorganize libmask to one module
The reason why this was done is that, since most modules have been
shrunken down to at most one item, it made sense to put them all in the
root module.

BREAKING CHANGE: The two structures and exec of libmask have moved to
the root module. exec has been renamed to haxe_exec.